+++ index.html
@@ -25,5 +25,6 @@
   </head>
   <body>
     <script type="module" src="solarSystem.js"></script>
+    <script type="module" src="solarSystemLoader.js"></script>
   </body>
 </html>
