+++ solarSystem.js
@@ -1,14 +1,13 @@
-﻿import * as THREE from "three"
+import * as THREE from "three"
 import { GUI } from "dat.gui"
 import { OrbitControls } from "three/examples/jsm/controls/OrbitControls"
 import { EffectComposer } from "three/examples/jsm/postprocessing/EffectComposer.js"
 import { RenderPass } from "three/examples/jsm/postprocessing/RenderPass.js"
 import { UnrealBloomPass } from "three/examples/jsm/postprocessing/UnrealBloomPass.js"
-import { GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader.js"
+import { loadSolarSystem } from "./solarSystemLoader.js"
 
 let scene, camera, renderer, composer, gui
 
-// Main: sets up renderer, scene, camera, controls, composer, and solar system components
 function main() {
   renderer = new THREE.WebGLRenderer()
   renderer.setClearColor(new THREE.Color(0x000000))
@@ -24,7 +23,7 @@ function main() {
 
   setupLighting()
 
-  createSolarSystem().then((solarSystem) => {
+  loadSolarSystem().then((solarSystem) => {
     scene.add(solarSystem)
   })
 
@@ -48,10 +47,10 @@ function initCamera() {
 
 // Initialize orbit controls and GUI
 function initControls() {
-  const orbitControls = new OrbitControls(camera, renderer.domElement)
-  const gui = new GUI()
-  const controls = { temp: 0 }
-  gui.add(controls, "temp", -10, 10).onChange(controls.redraw)
+  new OrbitControls(camera, renderer.domElement)
+  gui = new GUI()
+  const params = { temp: 0 }
+  gui.add(params, "temp", -10, 10)
   return gui
 }
 
@@ -80,83 +79,6 @@ function setupLighting() {
   // Add point light from the sun
   const pointLight = new THREE.PointLight(0xffffff, 100, 1000)
   scene.add(pointLight)
-}
-
-// Create the sun at the center of the solar system
-function createSun() {
-  const geometry = new THREE.SphereGeometry(5, 32, 32)
-  const material = new THREE.MeshBasicMaterial({ color: 0xffff00 })
-  const sun = new THREE.Mesh(geometry, material)
-  sun.position.set(0, 0, 0)
-  return sun
-}
-
-// Create the solar system with the sun and all planets
-async function createSolarSystem() {
-  const solarSystem = new THREE.Group()
-  const sun = createSun()
-  solarSystem.add(sun)
-
-  // Define the properties of each planet
-  const properties = [
-    { name: "mercury", distance: 10, scale: 2 },
-    { name: "venus", distance: 15, scale: 3 },
-    { name: "earth", distance: 22, scale: 3.5 },
-    { name: "mars", distance: 29, scale: 2.5 },
-    { name: "jupiter", distance: 40, scale: 6 },
-    { name: "saturn", distance: 50, scale: 5 },
-    { name: "uranus", distance: 60, scale: 4.5 },
-    { name: "neptune", distance: 70, scale: 4 },
-  ]
-
-  // Load each planet model and add it to the solar system
-  for (const property of properties) {
-    const planetGroup = new THREE.Group()
-    // Set the x position from the sun
-    planetGroup.position.set(property.distance, 0, 0)
-    try {
-      const planetModel = await loadPlanetModel(property.name, property.scale)
-      planetGroup.add(planetModel)
-    } catch (error) {
-      console.error(`Failed to load model for ${property.name}`, error)
-    }
-    solarSystem.add(planetGroup)
-  }
-
-  return solarSystem
-}
-
-// Load a planet model, normalize, and scale it
-function loadPlanetModel(planetName, desiredScale) {
-  return new Promise((resolve) => {
-    const loader = new GLTFLoader()
-    const url = `shader/${planetName}/scene.gltf`
-    loader.load(
-      url,
-      (gltf) => {
-        const model = gltf.scene
-        // Normalize and scale the model
-        normalizeModel(model)
-        model.scale.multiplyScalar(desiredScale)
-        resolve(model)
-      },
-      undefined,
-      (error) => {
-        console.log(`Error loading model for ${planetName}:`, error)
-      }
-    )
-  })
-}
-
-// Normalize a model so that it fits inside a unit bounding box
-function normalizeModel(object) {
-  const box = new THREE.Box3().setFromObject(object)
-  const size = new THREE.Vector3()
-  box.getSize(size)
-  const maxDim = Math.max(size.x, size.y, size.z)
-  if (maxDim === 0) return
-  const scale = 1 / maxDim
-  object.scale.set(scale, scale, scale)
 }
 
 // Creates a star field with 3000 stars
